+++ themes/dark/assets/js/templates/name-pattern.vue
@@ -1,5 +1,5 @@
 <template>
-    <div id="name_pattern_wrapper">
+    <div id="name-pattern-wrapper">
         <!-- If a 'type' is passed, it means that where checking a custom naming pattern. As for example sports, air-by-date etc.
         In that case, we're showing this checkbox, to display the rest of the form.
         If type evaulates to an empty string, we're asuming this is the default name pattern. And it's enabled by default. -->
@@ -8,7 +8,7 @@
                 <span>Custom {{ type }}</span>
             </label>
             <div class="col-sm-10 content">
-                <toggle-button :width="45" :height="22" id="enable_naming_custom" name="enable_naming_custom" v-model="isEnabled" sync></toggle-button>
+                <toggle-button :width="45" :height="22" id="enable_naming_custom" name="enable_naming_custom" v-model="isEnabled" @input="update()" sync></toggle-button>
                 <span>Name {{ type }} shows differently than regular shows?</span>
             </div>
         </div>
@@ -19,35 +19,24 @@
                     <span>Name Pattern:</span>
                 </label>
                 <div class="col-sm-10 content">
-                    <select id="name_presets" class="form-control input-sm" v-model="selectedNamingPattern" @change="updatePatternSamples">
-                        <option id="preset" v-for="preset in presets">{{ preset }}</option>
+                    <select id="name_presets" class="form-control input-sm" v-model="selectedNamingPattern" @change="updatePatternSamples" @input="update()">
+                        <option :id="preset.pattern" v-for="preset in presets" :key="preset.pattern">{{ preset.example }}</option>
                     </select>
                 </div>
             </div>
 
             <div id="naming_custom">
-                <div class="form-group" style="padding-top: 0;">
+                <div v-if="isCustom" class="form-group" style="padding-top: 0;">
                     <label class="col-sm-2 control-label">
                         <span>&nbsp;</span>
                     </label>
                     <div class="col-sm-10 content">
-                        <input v-if="isCustom" type="text" name="naming_pattern" id="naming_pattern" v-model="customName" @change="updatePatternSamples" class="form-control input-sm input350"/>
-                        <img src="images/legend16.png" width="16" height="16" alt="[Toggle Key]" id="show_naming_key" title="Toggle Naming Legend" class="legend" @click="showLegend = !showLegend" />
+                        <input type="text" name="naming_pattern" id="naming_pattern" v-model="customName" @change="updatePatternSamples" @input="update()" class="form-control-inline-max input-sm max-input350"/>
+                        <img src="images/legend16.png" width="16" height="16" alt="[Toggle Key]" id="show_naming_key" title="Toggle Naming Legend" class="legend" @click="showLegend = !showLegend"/>
                     </div>
                 </div>
 
-                <div v-if="selectedMultiEpStyle" class="form-group">
-                    <label class="col-sm-2 control-label" for="naming_multi_ep">
-                        <span>Multi-Episode Style:</span>
-                    </label>
-                    <div class="col-sm-10 content">
-                        <select id="naming_multi_ep" name="naming_multi_ep" v-model="selectedMultiEpStyle" class="form-control input-sm" @change="updatePatternSamples">
-                            <option id="multiEpStyle" :value="multiEpStyle.value" v-for="multiEpStyle in availableMultiEpStyles">{{ multiEpStyle.text }}</option>
-                        </select>
-                    </div>
-                </div>
-
-                <div id="naming_key" class="nocheck" v-if="showLegend">
+                <div id="naming_key" class="nocheck" v-if="showLegend && isCustom">
                     <table class="Key">
                         <thead>
                             <tr>
@@ -209,34 +198,42 @@
                             </tr>
                         </tbody>
                     </table>
-                    <br>
                 </div>
             </div>
 
-            <div id="naming_example_div">
-                <h3>Single-EP Sample:</h3>
-                <div class="example">
+            <div v-if="selectedMultiEpStyle" class="form-group">
+                <label class="col-sm-2 control-label" for="naming_multi_ep">
+                    <span>Multi-Episode Style:</span>
+                </label>
+                <div class="col-sm-10 content">
+                    <select id="naming_multi_ep" name="naming_multi_ep" v-model="selectedMultiEpStyle" class="form-control input-sm" @change="updatePatternSamples" @input="update($event)">
+                        <option id="multiEpStyle" :value="multiEpStyle.value" v-for="multiEpStyle in availableMultiEpStyles" :key="multiEpStyle.value">{{ multiEpStyle.text }}</option>
+                    </select>
+                </div>
+            </div>
+
+            <div class="form-group row">
+                <h3 class="col-sm-12">Single-EP Sample:</h3>
+                <div class="example col-sm-12">
                     <span class="jumbo" id="naming_example">{{ namingExample }}</span>
                 </div>
-                <br>
             </div>
-            <div v-if="isMulti" id="naming_example_multi_div">
-                <h3>Multi-EP sample:</h3>
-                <div class="example">
+            <div v-if="isMulti" class="form-group row">
+                <h3 class="col-sm-12">Multi-EP sample:</h3>
+                <div class="example col-sm-12">
                     <span class="jumbo" id="naming_example_multi">{{ namingExampleMulti }}</span>
                 </div>
-                <br>
             </div>
 
             <!-- Anime only -->
             <div v-if="animeType > 0" class="form-group">
                 <label for="naming_anime" class="col-sm-2 control-label">
                     <span>Add Absolute Number</span>
-                    <p>Add the absolute number to the season/episode format?</p>
                 </label>
                 <div class="col-sm-10 content">
-                    <input type="radio" name="naming_anime" id="naming_anime" value="1" v-model="animeType" @change="updatePatternSamples"/>
-                    <span>Only applies to animes. (eg. S15E45 - 310 vs S15E45)</span>
+                    <input type="radio" name="naming_anime" id="naming_anime" value="1" v-model="animeType" @change="updatePatternSamples" @input="update()" />
+                    <span>Add the absolute number to the season/episode format?</span>
+                    <p>Only applies to animes. (e.g. S15E45 - 310 vs S15E45)</p>
                 </div>
             </div>
 
@@ -245,7 +242,7 @@
                     <span>Only Absolute Number</span>
                 </label>
                 <div class="col-sm-10 content">
-                    <input type="radio" name="naming_anime" id="naming_anime_only" value="2" v-model="animeType" @change="updatePatternSamples"/>
+                    <input type="radio" name="naming_anime" id="naming_anime_only" value="2" v-model="animeType" @change="updatePatternSamples" @input="update()"/>
                     <span>Replace season/episode format with absolute number</span>
                     <p>Only applies to animes.</p>
                 </div>
@@ -256,28 +253,30 @@
                     <span>No Absolute Number</span>
                 </label>
                 <div class="col-sm-10 content">
-                    <input type="radio" name="naming_anime" id="naming_anime_none" value="3" v-model="animeType" @change="updatePatternSamples"/>
-                    <span>Only applies to animes.</span>
-                    <p>Dont include the absolute number</p>
+                    <input type="radio" name="naming_anime" id="naming_anime_none" value="3" v-model="animeType" @change="updatePatternSamples" @input="update()"/>
+                    <span>Don't include the absolute number</span>
+                    <p>Only applies to animes.</p>
                 </div>
             </div>
         </div>
+
     </div>
 </template>
+
 <script>
 module.exports = {
     name: 'name-pattern',
     props: {
         /**
          * Current naming pattern.
-        */
+         */
         namingPattern: {
             type: String,
             default: ''
         },
         /**
          * An array of available preset naming patterns.
-        */
+         */
         namingPresets: {
             type: Array,
             default: () => []
@@ -315,10 +314,14 @@ module.exports = {
         /**
          * Used icw with the type property.
          * If a type has been passed, the `enabled` property can be used to toggle the visibilty of the name-pattern settings.
-        */
+         */
         enabled: {
             type: Boolean,
             default: true
+        },
+        flagLoaded: {
+            type: Boolean,
+            default: false
         }
     },
     data() {
@@ -331,65 +334,94 @@ module.exports = {
             namingExample: '',
             namingExampleMulti: '',
             isEnabled: false,
-            isMulti: false,
             selectedMultiEpStyle: 1,
-            animeType: 0
-        }
+            animeType: 0,
+            lastSelectedPattern: ''
+        };
     },
     methods: {
         getDateFormat(format) {
             return dateFns.format(new Date(), format);
         },
         testNaming(pattern, selectedMultiEpStyle, animeType) {
-            console.debug('Test pattern ' + pattern + ' for ' + (selectedMultiEpStyle) ? 'multi' : 'single' + ' ep');
+            console.debug(`Test pattern ${pattern} for ${(selectedMultiEpStyle) ? 'multi' : 'single ep'}`);
             const params = {
                 pattern,
-                anime_type: animeType
-            }
+                anime_type: animeType // eslint-disable-line camelcase
+            };
 
             if (selectedMultiEpStyle) {
                 params.multi = selectedMultiEpStyle;
             }
 
             try {
-                return apiRoute.get('config/postProcessing/testNaming', {params: params}).then(res => res.data);
-            } catch (e) {
-                console.warn(e);
+                return apiRoute.get('config/postProcessing/testNaming', { params, timeout: 20000 }).then(res => res.data);
+            } catch (error) {
+                console.warn(error);
                 return '';
             }
         },
-        async updatePatternSamples() {
+        updatePatternSamples() {
             // If it's a custom pattern, we need to get the custom pattern from this.customName
+            // if customName if empty for whatever reason, just use the last selected preset value.
+            if (!this.customName) {
+                this.customName = this.lastSelectedPattern;
+            }
+
             const pattern = this.isCustom ? this.customName : this.pattern;
 
+            // Exiting early as we probably don't have all the properties yet.
+            // updatePatternSamples() can be triggered from a watcher on pattern, this.selectedMultiEpStyle or this.animeType
+            // We want to make sure that the data passed on to the component is complete before making calls to the backend.
+            // If we don't check on this, it will send api requests with null data.
+
+            if (!pattern || this.animeType === null || this.selectedMultiEpStyle === null) {
+                return;
+            }
+
             // Update single
-            this.namingExample = await this.testNaming(pattern, false, this.animeType) + '.ext';
+            this.testNaming(pattern, false, this.animeType).then(result => {
+                this.namingExample = result + '.ext';
+            });
+
             console.debug('Result of naming pattern check: ' + this.namingExample);
 
             // Test naming
             this.checkNaming(pattern, false, this.animeType);
 
-            // Update multi if needed
             if (this.isMulti) {
-                this.namingExampleMulti = await this.testNaming(pattern, this.selectedMultiEpStyle, this.animeType) + '.ext';
+                this.testNaming(pattern, this.selectedMultiEpStyle, this.animeType).then(result => {
+                    this.namingExampleMulti = result + '.ext';
+                });
+
                 this.checkNaming(pattern, this.selectedMultiEpStyle, this.animeType);
             }
         },
-        update(newValue) {
-            this.$emit('change', {
-                pattern: this.isCustom ? this.customName : this.pattern,
-                type: this.type,
-                multiEpStyle: this.selectedMultiEpStyle,
-                custom: this.isCustom,
-                enabled: this.isEnabled,
-                animeNamingType: this.animeType
+        update() {
+            if (!this.flagLoaded) {
+                return;
+            }
+
+            this.$nextTick(() => {
+                this.$emit('change', {
+                    pattern: this.isCustom ? this.customName : this.pattern,
+                    type: this.type,
+                    multiEpStyle: this.selectedMultiEpStyle,
+                    custom: this.isCustom,
+                    enabled: this.isEnabled,
+                    animeNamingType: Number(this.animeType)
+                });
             });
         },
         checkNaming(pattern, selectedMultiEpStyle, animeType) {
+            if (!pattern) {
+                return;
+            }
+
             const params = {
                 pattern,
-                anime_type: animeType
-            }
+                anime_type: animeType // eslint-disable-line camelcase
+            };
 
             if (selectedMultiEpStyle) {
                 params.multi = selectedMultiEpStyle;
@@ -398,7 +430,7 @@ module.exports = {
             const { $el } = this;
             const el = $($el);
 
-            const result = apiRoute.get('config/postProcessing/isNamingValid', { params }).then(res => {
+            apiRoute.get('config/postProcessing/isNamingValid', { params, timeout: 20000 }).then(result => {
                 if (result.data === 'invalid') {
                     el.find('#naming_pattern').qtip('option', {
                         'content.text': 'This pattern is invalid.',
@@ -424,37 +456,63 @@ module.exports = {
             }).catch(error => {
                 console.warn(error);
             });
+        },
+        updateCustomName() {
+            // Store the custom naming pattern.
+            if (!this.presetsPatterns.includes(this.pattern)) {
+                this.customName = this.pattern;
+            }
+
+            // If the custom name is empty, let's use the last selected pattern.
+            // We'd prefer to cache the last configured custom pattern.
+            if (!this.customName) {
+                this.customName = this.lastSelectedPattern;
+            }
         }
     },
     computed: {
         isCustom() {
-            return !this.presets.includes(this.pattern) || this.pattern === 'Custom...';
+            if (this.pattern) {
+                return !this.presetsPatterns.includes(this.pattern) || this.pattern === 'Custom...';
+            }
+            return false;
         },
         selectedNamingPattern: {
-            get: function() {
-                return this.isCustom ? 'Custom...' : this.pattern;
+            get() {
+                const filterPattern = () => {
+                    const foundPattern = this.presets.filter(preset => preset.pattern === this.pattern);
+                    if (foundPattern.length > 0) {
+                        return foundPattern[0].example;
+                    }
+                    return false;
+                };
+                return this.isCustom ? 'Custom...' : filterPattern();
             },
-            set: function(value) {
-                this.pattern = value;
+            set(example) {
+                // We need to convert the selected example back to a pattern
+                this.pattern = this.presets.filter(preset => preset.example === example)[0].pattern;
             }
+        },
+        presetsPatterns() {
+            return this.presets.map(preset => preset.pattern);
+        },
+        isMulti() {
+            return Boolean(this.multiEpStyle);
         }
     },
     mounted() {
         this.pattern = this.namingPattern;
 
-        // Store the custom naming pattern.
-        if (!this.presets.includes(this.pattern)) {
-            this.customName = this.pattern;
-        }
-
         // Add Custom... as an option to the presets.
-        this.presets = this.namingPresets.concat('Custom...');
+        this.presets = this.namingPresets.concat({ pattern: 'Custom...', example: 'Custom...' });
+
+        // Update the custom name
+        this.updateCustomName();
 
         // Pass properties into local variables
         this.availableMultiEpStyles = this.multiEpStyles;
         this.selectedMultiEpStyle = this.multiEpStyle;
         this.animeType = this.animeNamingType;
-        this.isMulti = Boolean(this.multiEpStyle)
 
         // If type is falsy, we asume it's the default name pattern. And thus enabled by default.
         this.isEnabled = this.type ? false : this.enabled;
@@ -463,20 +521,33 @@ module.exports = {
         this.updatePatternSamples();
     },
     watch: {
-        customName(newValue, oldValue) {
-            this.update(newValue);
+        // Update local variables when properties are updated
+        enabled() {
+            this.isEnabled = this.enabled;
         },
-        pattern(newValue, oldValue) {
-            this.update();
+        namingPattern(newPattern, oldPattern) {
+            this.lastSelectedPattern = newPattern || oldPattern;
+
+            this.pattern = this.namingPattern;
+            this.updateCustomName();
+            this.updatePatternSamples();
         },
-        selectedMultiEpStyle(newValue) {
-            this.update();
+        namingPresets() {
+            this.presets = this.namingPresets;
         },
-        isEnabled() {
-            this.update();
+        multiEpStyle() {
+            this.selectedMultiEpStyle = this.multiEpStyle;
+            this.updatePatternSamples();
         },
-        animeType() {
-            this.update();
+        multiEpStyles() {
+            this.availableMultiEpStyles = this.multiEpStyles;
+        },
+        animeNamingType() {
+            this.animeType = this.animeNamingType;
+            this.updatePatternSamples();
+        },
+        type() {
+            this.isEnabled = this.type ? false : this.enabled;
         }
     }
 };